Comprehensive Classification Notebooks
These Notebooks are divided into three subsequent stages. Part 1 deals with the first inspection of the data and the cleansing of the data set. Part 2 delves deeper into the subject by an in-depth analysis. The actual classification starts with Part 3. This part is split into 3.1 (k-Nearest Neighbor Classifier) and 3.2 (Random Forest Classifier).